### Purpose:

make the trade store use significantly less CPU, and make the test run a bit faster (but they seem to be mostly I/O bound anyway).

### Current Behavior:

In order to convert between `TradeRecord` and `TradeRecordOld`, we go via `dataclasses.asdict()`, which performs a deep copy of the whole class. This is expensive, see profile below.

### New Behavior:

Converting between `TradeRecord` and `TradeRecordOld` is simply done by passing in the fields (by reference).


### Testing Notes:

| function | before | after | after/before |
| --- | --- | --- | --- |
| add_trade_record | 33.01% | 26.53 % | 80.37% |
| get_all_trades | 11.62% | 4.41% | 37.95% |

The run time of the test is not materially affected:

```
before: 435.29s (0:07:15)
after: 423.02s (0:07:03)
```

#### before

![before](https://github.com/Chia-Network/chia-blockchain/assets/661450/3f321896-9d11-4458-8a55-650caf89cf8c)


#### after

![after](https://github.com/Chia-Network/chia-blockchain/assets/661450/58c949bf-9134-4c8c-932c-088275fcac21)